Refactoring: cloud functions
